crates/rout3serv_py/pyproject.toml
[tool.poetry]
name = "rout3serv"
version = "0.2.0"
description = "api client for rout3serv"
authors = ["Nico Mandery <nico@nmandery.net>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.38.0"
Shapely = "^1.7.1"
pyarrow = "^6.0.0"
protobuf = "^3.19.1"
#pandas = "^1.3.0"
#geopandas = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
grpcio-tools = "^1.38.0"
notebook = "^6.4.0"
geopandas = "^0.9.0"
folium = "^0.12.1"
h3 = "^3.7.3"
branca = "^0.4.2"
geojson = "^2.5.0"
matplotlib = "^3.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
